More duplicate code removal
