+++ imports/api/company/companyMethods.js
@@ -3,18 +3,20 @@ import { Mongo } from 'meteor/mongo';
 import validate from 'validate.js';
 
 import Companies from './companyCollection';
-import { properties } from '../../components/pages/company/Company';
 import { buildSearchRegExp } from '../searchUtils';
-
-const CREATION = 'CREATION';
-const NOTE = 'NOTE';
+import FieldLists from '../fieldList/fieldListCollection';
 
 import * as GM from '../genericMethods';
+
+const properties = () =>
+  FieldLists.findOne({ page: 'COMPANY_PROPERTIES' })
+    ? FieldLists.findOne({ page: 'COMPANY_PROPERTIES' }).fields
+    : [];
 
 // Generic Methods
 export const create = company => GM.create(Companies, company);
 export const saveProperties = (companyId, company) =>
-  GM.saveProperties(Companies, properties, companyId, company);
+  GM.saveProperties(Companies, properties(), companyId, company);
 export const addNote = (companyId, note) =>
   GM.addNote(Companies, companyId, note);
