+++ imports/api/contact/contactMethods.js
@@ -3,15 +3,20 @@ import { Mongo } from 'meteor/mongo';
 import validate from 'validate.js';
 
 import Contacts from './contactCollection';
-import { properties } from '../../components/pages/contact/Contact';
 import { buildSearchRegExp } from '../searchUtils';
+import FieldLists from '../fieldList/fieldListCollection';
 
 import * as GM from '../genericMethods';
+
+const properties = () =>
+  FieldLists.findOne({ page: 'CONTACT_PROPERTIES' })
+    ? FieldLists.findOne({ page: 'CONTACT_PROPERTIES' }).fields
+    : [];
 
 // Generic Methods
 export const create = contact => GM.create(Contacts, contact);
 export const saveProperties = (contactId, contact) =>
-  GM.saveProperties(Contacts, properties, contactId, contact);
+  GM.saveProperties(Contacts, properties(), contactId, contact);
 export const addNote = (contactId, note) =>
   GM.addNote(Contacts, contactId, note);
